refactor : hooks/redux를 사용해서 로직과 상태를 관리함
