+++ client/src/App.js
@@ -2,7 +2,7 @@ import './App.scss';
 import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
 
 import LandingPage from './components/LandingPage';
-import MapPage from './components/MapPage/Container';
+import MapPage from './components/MapPage';
 
 function App() {
   return (
