+++ client/src/modules/index.js
@@ -1,6 +1,7 @@
 import { combineReducers } from 'redux';
 import { all, fork } from 'redux-saga/effects';
 import { newSnackSaga, newSnack } from './newSnack';
+import { fullMapSaga, fullMap } from './fullMap';
 
 export const rootReducer = combineReducers({ newSnack });
